more implementations on issue #93

+++ src/BoxBack.Domain/Models/RotinaEventHistory.cs
@@ -27,8 +27,8 @@ namespace BoxBack.Domain.Models
         public DateTimeOffset DataInicio { get; set; }
         public DateTimeOffset DataFim { get; set; }
         public RotinaStatusProgressoEnum StatusProgresso { get; set; }
-        public Int32 TotalItensSucesso { get; set;}
-        public Int32 TotalItensInsucesso { get; set;}
+        public Int64 TotalItensSucesso { get; set;}
+        public Int64 TotalItensInsucesso { get; set;}
         public string ExceptionMensagem { get; set; }
